Improve simd search (25% improvement)
Replace semi_at and next_newline with a simpler (and seemingly
significantly faster) simd search.

I define a `find_byte` function which is generic over the number of SIMD
lanes. I then select lane sizes for the newline and semicolon searches that minimize total time
(on my machine anyway) which wound up being significantly smaller than the previous choice of 64.

I suspect this is because newlines and (especially) semicolons are often
found relatively early and using a smaller number of lanes improves
latency (at the hypothetical cost of throughput).

This also removes the need for `memchr` (calling it on the
slow path yields no discernable difference).